Refactor into seperate js files.
